requirements.txt
torch
streamlit
pillow
torchvision
huggingface-hub
tqdm
nltk
